Added Eslint and Prettier configurations
