+++ package.json
@@ -4,6 +4,8 @@
     "ios": "react-native run-ios",
     "web": "expo start --web",
     "start": "react-native start",
+    "format": "prettier \"**/*.{js,html}\" --write",
+    "lint": "eslint \"**/*.{js,jsx}\" --quiet",
     "test": "jest"
   },
   "dependencies": {
@@ -11,9 +13,9 @@
     "@react-navigation/bottom-tabs": "^5.1.0",
     "@react-navigation/native": "^5.0.8",
     "expo": "~36.0.0",
-    "react": "~16.9.0",
-    "react-dom": "~16.9.0",
-    "react-native": "~0.61.4",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
+    "react-native": "^0.61.5",
     "react-native-gesture-handler": "^1.6.0",
     "react-native-reanimated": "^1.7.0",
     "react-native-safe-area-context": "^0.7.3",
@@ -23,9 +25,16 @@
   },
   "devDependencies": {
     "@babel/core": "~7.6.0",
+    "babel-eslint": "^10.1.0",
     "babel-jest": "~24.9.0",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.10.0",
+    "eslint-plugin-import": "^2.20.1",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.18.3",
     "jest": "~24.9.0",
     "metro-react-native-babel-preset": "~0.56.0",
+    "prettier": "^1.19.1",
     "react-test-renderer": "~16.9.0"
   },
   "jest": {
